+++ packages/router/src/adapters/subgraph/subgraph.ts
@@ -103,7 +103,7 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
 
   const { requestContext, methodContext } = createLoggingContext(getActiveTransactions.name, _requestContext);
 
-  const routerAddress = wallet.address;
+  const routerAddress = await wallet.getAddress();
 
   // get local context
   const sdks = getSdks();
@@ -135,7 +135,7 @@ export const getActiveTransactions = async (_requestContext?: RequestContext): P
             allReceiverExpired: jsonifyError(allReceiverExpired as any),
           });
         }
-        
+
         // get all sender prepared txs
         const allSenderPrepared = await sdk.request<GetSenderTransactionsQuery>((client) =>
           client.GetSenderTransactions({
@@ -361,7 +361,7 @@ export const getTransactionForChain = async (
   const methodId = getUuid();
 
   const { wallet } = getContext();
-  const routerAddress = wallet.address;
+  const routerAddress = await wallet.getAddress();
 
   const sdks = getSdks();
   const sdk = sdks[chainId];
@@ -421,7 +421,8 @@ export const getAssetBalance = async (assetId: string, chainId: number): Promise
     throw new ContractReaderNotAvailableForChain(chainId, { method, methodId });
   }
 
-  const assetBalanceId = `${assetId.toLowerCase()}-${wallet.address.toLowerCase()}`;
+  const routerAddress = await wallet.getAddress();
+  const assetBalanceId = `${assetId.toLowerCase()}-${routerAddress.toLowerCase()}`;
   const bal = await sdk.request<GetAssetBalanceQuery>((client) => client.GetAssetBalance({ assetBalanceId }));
   return bal.assetBalance?.amount ? BigNumber.from(bal.assetBalance?.amount) : constants.Zero;
 };